refactor heredocumet
