+++ srcs/heredocmet/readline_heredoc.c
@@ -1,0 +1,57 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   readline_heredoc.c                                 :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: jnakahod <jnakahod@student.42tokyo.jp>     +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2021/10/16 12:33:54 by jnakahod          #+#    #+#             */
+/*   Updated: 2021/10/16 12:58:02 by jnakahod         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include <mini_shell.h>
+
+static char	*update_heredoc(char **old, char *input)
+{
+	char	*joined_input;
+	char	*new_heredoc;
+
+	joined_input = NULL;
+	new_heredoc = NULL;
+	joined_input = ft_strjoin(*old, input);
+	free_set((void **)old, NULL);
+	if (!joined_input)
+		return (NULL);
+	new_heredoc = ft_strjoin(joined_input, "\n");
+	free_set((void **)&joined_input, NULL);
+	if (!new_heredoc)
+		return (NULL);
+	return (new_heredoc);
+}
+
+static t_bool	can_exit_heredocument(char *input, char *delimiter)
+{
+	if (!input || is_match_str(input, delimiter))
+		return (TRUE);
+	else
+		return (FALSE);
+}
+
+t_result	readline_input_heredoc(char **heredoc, char *delimiter)
+{
+	char	*input;
+
+	input = NULL;
+	free_set((void **)heredoc, NULL);
+	while (TRUE)
+	{
+		input = readline("> ");
+		if (can_exit_heredocument(input, delimiter))
+			break ;
+		*heredoc = update_heredoc(heredoc, input);
+		free_set((void **)&input, NULL);
+	}
+	free_set((void **)&input, NULL);
+	return (SUCCESS);
+}
